[ISSUE-2905] Adding Shuffle Giphy extension
